svn merge file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@2327 file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@HEAD .
git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/branches/pdonnel3/search-syscall@2659 a4d8336d-3463-0410-8bba-c098c45d37a8

+++ chirp/src/chirp_distribute.c
@@ -293,7 +293,7 @@ int main(int argc, char *argv[])
 
 	result = chirp_reli_stat(sourcehost, sourcepath, &buf, time(0) + 20);
 	if(result < 0) {
-		printf("%s %s %lld %i %s chirp stat failed\n", sourcehost, sourcepath, result, errno, strerror(errno));
+		printf("%s %s %" PRId64 " %i %s chirp stat failed\n", sourcehost, sourcepath, result, errno, strerror(errno));
 		if(errno == 2)
 			return 0;
 	}
@@ -514,9 +514,6 @@ int main(int argc, char *argv[])
 	if(out != NULL)
 		fclose(out);
 
-
-
-
 	targets = malloc(sizeof(struct target_info) * ntargets);
 	servers = malloc(sizeof(struct server_info) * 400);
 
@@ -541,7 +538,7 @@ int main(int argc, char *argv[])
 
 			if(strlen(temp) > 10)
 				cluster_count = cluster_count + 1;
-			for(i = 0; i < strlen(temp); i++) {
+			for(i = 0; i < (int) strlen(temp); i++) {
 				if((test != 0) && (temp[i] != '_') && (temp[i] != ' ')) {
 					t_name[name_count] = temp[i];
 					name_count++;
